excercisenine

+++ exercise-nine/src/ar/com/ada/online/second/exercisenine/Main.java
@@ -6,20 +6,22 @@ public class Main {
 
     public static void main(String[] args) {
 
-
+        // Se requiere determinar el costo que tendrá realizar una llamada telefónica con base en
+        //el tiempo que dura la llamada y en el costo por minuto. Realice un diagrama de flujo y
+        //pseudocódigo que representen el algoritmo para tal fin.
         System.out.println("Programa que determina el costo de una llamada telefonica");
 
         // Defino las variables
-        int callCost;
+        double callCost;
         int callDuration;
-        int costPerMinute;
+        double costPerMinute;
 
         // Habilito el ingreso de datos por el teclado
         Scanner keyboard = new Scanner(System.in);
 
         // Se leen los datos que se ingresan
         System.out.print("Ingrese el costo por minuto:$ ");
-        costPerMinute = keyboard.nextInt();
+        costPerMinute = keyboard.nextDouble();
 
         System.out.print("Ingrese los minutos que duró la llamada: ");
         callDuration = keyboard.nextInt();
